refactor: rename product_service app to products
